+++ com.unity.cinemachine/Editor/SaveDuringPlay/SaveDuringPlay.cs
@@ -5,6 +5,7 @@ using System.Reflection;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.Assertions;
+using UnityEngine.SceneManagement;
 
 namespace Unity.Cinemachine.Editor
 {
@@ -26,7 +27,7 @@ namespace Unity.Cinemachine.Editor
         /// <summary>
         /// Will find the named object, active or inactive, from the full path.
         /// </summary>
-        public static GameObject FindObjectFromFullName(string fullName, GameObject[] roots)
+        public static GameObject FindObjectFromFullName(string fullName, List<GameObject> roots)
         {
             if (string.IsNullOrEmpty(fullName) || roots == null)
                 return null;
@@ -36,7 +37,7 @@ namespace Unity.Cinemachine.Editor
                 return null;
 
             Transform root = null;
-            for (int i = 0; root == null && i < roots.Length; ++i)
+            for (int i = 0; root == null && i < roots.Count; ++i)
                 if (roots[i].name == path[1])
                     root = roots[i].transform;
 
@@ -62,17 +63,20 @@ namespace Unity.Cinemachine.Editor
             return root.gameObject;
         }
 
-        /// <summary>Finds all the root objects in a scene, active or not</summary>
-        public static GameObject[] FindAllRootObjectsInScene()
+        /// <summary>Finds all the root objects, active or not, in all open scenes</summary>
+        public static List<GameObject> FindAllRootObjectsInOpenScenes()
         {
-            return UnityEngine.SceneManagement.SceneManager.GetActiveScene().GetRootGameObjects();
+            var allRoots = new List<GameObject>();
+            for (int i = 0; i < SceneManager.sceneCount; ++i)
+                allRoots.AddRange(SceneManager.GetSceneAt(i).GetRootGameObjects());
+            return allRoots;
         }
 
 
         /// <summary>
-        /// This finds all the behaviours in scene, active or inactive, excluding prefabs
+        /// This finds all the behaviours, active or inactive, in open scenes, excluding prefabs
         /// </summary>
-        public static T[] FindAllBehavioursInScene<T>() where T : MonoBehaviour
+        public static List<T> FindAllBehavioursInOpenScenes<T>() where T : MonoBehaviour
         {
             List<T> objectsInScene = new ();
             var allObjects = Resources.FindObjectsOfTypeAll<T>();
@@ -88,7 +92,7 @@ namespace Unity.Cinemachine.Editor
                     continue;
                 objectsInScene.Add(b);
             }
-            return objectsInScene.ToArray();
+            return objectsInScene;
         }
     }
 
@@ -302,6 +306,8 @@ namespace Unity.Cinemachine.Editor
 
         Dictionary<string, string> mValues = new Dictionary<string, string>();
 
+        public string ObjetFullPath => mObjectFullPath;
+
         /// <summary>
         /// Recursively collect all the field values in the MonoBehaviours
         /// owned by this object and its descendants.  The values are stored
@@ -310,7 +316,7 @@ namespace Unity.Cinemachine.Editor
         public void CollectFieldValues(GameObject go)
         {
             mObjectFullPath = ObjectTreeUtil.GetFullName(go);
-            GameObjectFieldScanner scanner = new GameObjectFieldScanner();
+            GameObjectFieldScanner scanner = new ();
             scanner.FilterField = FilterField;
             scanner.FilterComponent = HasSaveDuringPlay;
             scanner.OnLeafField = (string fullName, Type type, ref object value) =>
@@ -323,7 +329,7 @@ namespace Unity.Cinemachine.Editor
             scanner.ScanFields(go);
         }
 
-        public GameObject FindSavedGameObject(GameObject[] roots)
+        public GameObject FindSavedGameObject(List<GameObject> roots)
         {
             return ObjectTreeUtil.FindObjectFromFullName(mObjectFullPath, roots);
         }
@@ -335,7 +341,7 @@ namespace Unity.Cinemachine.Editor
         /// value in the game object, Set the GameObject's value using the value
         /// recorded in the dictionary.
         /// </summary>
-        public bool PutFieldValues(GameObject go, GameObject[] roots)
+        public bool PutFieldValues(GameObject go, List<GameObject> roots)
         {
             GameObjectFieldScanner scanner = new GameObjectFieldScanner();
             scanner.FilterField = FilterField;
@@ -389,7 +395,7 @@ namespace Unity.Cinemachine.Editor
         /// because the reflection system breaks them down into their primitive components.
         /// You can add more support here, as needed.
         /// </summary>
-        static object LeafObjectFromString(Type type, string value, GameObject[] roots)
+        static object LeafObjectFromString(Type type, string value, List<GameObject> roots)
         {
             if (type == typeof(Single))
                 return float.Parse(value);
@@ -520,8 +526,8 @@ namespace Unity.Cinemachine.Editor
         static HashSet<GameObject> FindInterestingObjects()
         {
             var objects = new HashSet<GameObject>();
-            var everything = ObjectTreeUtil.FindAllBehavioursInScene<MonoBehaviour>();
-            for (int i = 0; i < everything.Length; ++i)
+            var everything = ObjectTreeUtil.FindAllBehavioursInOpenScenes<MonoBehaviour>();
+            for (int i = 0; i < everything.Count; ++i)
             {
                 var b = everything[i];
                 if (!objects.Contains(b.gameObject) && ObjectStateSaver.HasSaveDuringPlay(b))
@@ -557,7 +563,7 @@ namespace Unity.Cinemachine.Editor
         {
             //Debug.Log("Updating state for all interesting objects");
             bool dirty = false;
-            var roots = ObjectTreeUtil.FindAllRootObjectsInScene();
+            var roots = ObjectTreeUtil.FindAllRootObjectsInOpenScenes();
             for (int i = 0; i < s_SavedStates.Count; ++i)
             {
                 var saver = s_SavedStates[i];